Moving encoding method to SlowCheetah class Adding method for stream encoding
