+++ src/Microsoft.VisualStudio.SlowCheetah/Transformer/TransformUtilities.cs
@@ -1,0 +1,69 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.VisualStudio.SlowCheetah
+{
+    using System.IO;
+    using System.Text;
+
+    /// <summary>
+    /// Utilities for transformations
+    /// </summary>
+    public static class TransformUtilities
+    {
+        /// <summary>
+        /// Determines a text file's encoding by analyzing its byte order mark (BOM).
+        /// Defaults to ASCII when detection of the text file's endianness fails.
+        /// </summary>
+        /// <param name="filename">The text file to analyze.</param>
+        /// <returns>The detected encoding.</returns>
+        public static Encoding GetEncoding(string filename)
+        {
+            using (var file = new FileStream(filename, FileMode.Open, FileAccess.Read))
+            {
+                return GetEncoding(file);
+            }
+        }
+
+        /// <summary>
+        /// Determines a stream's encoding by analyzing its byte order mark (BOM).
+        /// Defaults to ASCII when detection of the text file's endianness fails.
+        /// </summary>
+        /// <param name="stream">The stream to analyze.</param>
+        /// <returns>The detected encoding.</returns>
+        public static Encoding GetEncoding(Stream stream)
+        {
+            // Read the BOM
+            var bom = new byte[4];
+            stream.Read(bom, 0, 4);
+
+            // Analyze the BOM
+            if (bom[0] == 0x2b && bom[1] == 0x2f && bom[2] == 0x76)
+            {
+                return Encoding.UTF7;
+            }
+
+            if (bom[0] == 0xef && bom[1] == 0xbb && bom[2] == 0xbf)
+            {
+                return Encoding.UTF8;
+            }
+
+            if (bom[0] == 0xff && bom[1] == 0xfe)
+            {
+                return Encoding.Unicode; // UTF-16LE
+            }
+
+            if (bom[0] == 0xfe && bom[1] == 0xff)
+            {
+                return Encoding.BigEndianUnicode; // UTF-16BE
+            }
+
+            if (bom[0] == 0 && bom[1] == 0 && bom[2] == 0xfe && bom[3] == 0xff)
+            {
+                return Encoding.UTF32;
+            }
+
+            return Encoding.ASCII;
+        }
+    }
+}
